wow ... this is so damn subtle it is not even funny.
if you have max-height set when codemirror creates a hidden div to measure the size of the area so it can render the cursor something odd happens ...

it creates a div with 50 x's ... well the max-height stuff messes it up. If the height is restricted and codemirror thinks the panel it is working is 600px as opposed to the real height.

end result ... weird size cursors at the wrong position